+++ parsley.js
@@ -498,7 +498,7 @@
     * @method manageErrorContainer
     */
     , manageErrorContainer: function () {
-      var errorContainer = this.options.errorContainer || this.options.errors.container( this.element, this.ParsleyInstance.isRadioOrCheckbox )
+      var errorContainer = this.options.errorContainer || this.options.errors.container( this.ParsleyInstance.element, this.ParsleyInstance.isRadioOrCheckbox )
         , ulTemplate = this.options.animate ? this.ulTemplate.css('display', '') : this.ulTemplate;
 
       if ( 'undefined' !== typeof errorContainer ) {